% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcovidplots-package.R
\name{drcovidplots}
\alias{drcovidplots}
\title{package-drcovidplots}
\description{
: This packages generates several plots using COVID-19 data from
the Minister of Public Health and Social Assintance from Dominican Republic for COVID19.
}
\section{Warning}{

Before using any function, you need to run,
in order, \code{load_data_covid_dr()} and \code{load_themes()}
}

\keyword{"drcovidplots"}
\keyword{internal}
