% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_active_cases.R
\name{g_active_cases}
\alias{g_active_cases}
\title{Active cases of COVID19}
\usage{
g_active_cases(saveplot = FALSE, savepng = FALSE)
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default \code{FALSE}.}

\item{savepng}{Logical. Should save a png version of the plot? Default \code{FALSE}.}
}
\value{
Chart of COVID19 active cases and save a copy in png format on the computer
at the address defined in \code{setwd()}.
}
\description{
This function graphs the actives cases cases of COVID19 in the DR.
}
\examples{
g_active_cases(saveplot = FALSE, savepng = TRUE)
}
