% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_cases_province.R
\name{g_cases_province}
\alias{g_cases_province}
\title{Provinces with most positive cases of COVID19}
\usage{
g_case_province()
}
\arguments{
\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph of the provinces with the most positive cases of COVID19 and saves a
copy in png and gif format to the computer at the address defined in \code{setwd()}.
}
\description{
This function creategraphs the provinces with the most positive cases of COVID19 in the DR.
}
\examples{
g_case_provinces()
g_case_provinces(savepng = TRUE)
}
