% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_death.R
\name{g_death}
\alias{g_death}
\title{Total number of deaths}
\usage{
g_death(saveplot = FALSE, savepng = FALSE)
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default \code{FALSE}.}

\item{savepng}{Logical. Should save a png version of the plot? Default \code{FALSE}.}
}
\value{
Graph of the total number of deaths and saves a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs the total number of deaths by COVID19 in the Dominican Republic.
}
\examples{
g_death()
g_death(savepng = TRUE)
}
