% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_density.R
\name{g_density}
\alias{g_density}
\title{Correlation between COVID19 cases and population density}
\usage{
g_density()
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default FALSE.}

\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph of daily and accumulated tests both in units and per million inhabitants
and save a copy in png format on the computer to the address defined in \code{setwd()}.
}
\description{
This function graphs the correlation between cases of COVID19
and population density of COVID19 in DR.
}
\examples{
g_density()
}
