% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_evolution_covid.R
\name{g_evolution_covid}
\alias{g_evolution_covid}
\title{Evolution of new cases and total cases of COVID19}
\usage{
g_evolution_covid()
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default FALSE.}

\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
GIF with the evolution of the new cases and total cases and save a
copy to the computer at the address defined in \code{setwd()}.
}
\description{
This function generates a gif with the evolution of the new cases and total cases of COVID19 in DR.
}
\examples{
g_evolution_covid()
g_evolution_covid(saveplot = TRUE)
}
