% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_growth_cases.R
\name{g_growth_cases}
\alias{g_growth_cases}
\title{Growth rate of positive cases of COVID19}
\usage{
g_growth_cases(saveplot = FALSE, savepng = FALSE)
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default FALSE.}

\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph of the growth rate of positive cases and save a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs the growth rate of positive cases of COVID19 in DR.
}
\examples{
g_growth_cases(saveplot = TRUE, savepng = FALSE)
}
