% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_hospital.R
\name{g_hospital}
\alias{g_hospital}
\title{People in hospital isolation for COVID19}
\usage{
g_hospital(saveplot = FALSE, savepng = FALSE)
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default \code{FALSE}.}

\item{savepng}{Logical. Should save a png version of the plot? Default \code{FALSE}.}
}
\value{
Chart of isolated individuals in hospitals and save a copy in png format on the computer
at the address defined in \code{setwd()}.
}
\description{
This function graphs the people in hospital isolation of the positive cases of COVID19 in the DR.
}
\examples{
g_hospital(saveplot = FALSE, savepng = TRUE)
}
