% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_incidence_region.R
\name{g_incidence_region}
\alias{g_incidence_region}
\title{Incidence and lethality rate by region of COVID19}
\usage{
g_incidence_region()
}
\arguments{
\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph of incidence and case fatality by region and save a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs the incidence and lethality by region of the COVID19 in the DR.
}
\examples{
g_incidence_region()
}
