% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_log_nc.R
\name{g_log_nc}
\alias{g_log_nc}
\title{New cases vs total cases in logarithms of COVID19}
\usage{
g_log_nc()
}
\arguments{
\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph of the logarithm of of both new cases and total cases and save a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs the logarithm of both new cases and total cases
in the Dominican Republic
}
\examples{
g_log_nc()
}
