% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_map_covid.R
\name{g_map_covid}
\alias{g_map_covid}
\title{Dominican map with COVID-19 cases by province}
\usage{
g_map_covid(date = "latest", interactive = FALSE, variable = "Cases")
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default FALSE.}

\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph the Dominican map with COVID-19 cases by province and saves a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs the provinces with the most positive
  cases of COVID19 in the DR.
}
\examples{
g_map_covid()
g_map_covid(savepng = TRUE)
g_map_covid(date = "latest", variable = "Deaths")
}
