% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_np.R
\name{g_np}
\alias{g_np}
\title{Negative and positive cases of COVID19}
\usage{
g_np(saveplot = FALSE, savepng = FALSE)
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default FALSE.}

\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph of the total number of negative and positives cases and saves a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs the total number negative and positive COVID19 cases
in the Dominican Republic.
}
\examples{
g_np()
g_np(saveplot = FALSE, savepng = TRUE)
}
