% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_positive_region.R
\name{g_positive_region}
\alias{g_positive_region}
\title{Positiveness by region from COVID19 in the Dominican Republic}
\usage{
g_positive_region()
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default FALSE.}

\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph of the the positive ratio by region and saves a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs the positive ratio (Positive cases / Tests)
from COVID19 in the Dominican Republic ordered by provinces.
}
\examples{
g_positive_region()
g_positive_region(saveplot = FALSE, savepng = FALSE)
}
