% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_recovered.R
\name{g_recovered}
\alias{g_recovered}
\title{Recovered from COVID19}
\usage{
g_recovered()
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default FALSE.}

\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph of the recovered persons and saves a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs those recovered from COVID19 in the Dominican Republic.
}
\examples{
g_recovered()
g_recovered(saveplot = FALSE, savepng = FALSE)
}
