% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_recovered_daily.R
\name{g_recovered_daily}
\alias{g_recovered_daily}
\title{Daily number of persons recovered from COVID19}
\usage{
g_recovered_daily()
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default FALSE.}

\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph of the daily recovered persons and saves a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs the daily number of persons recovered from COVID19
in the Dominican Republic.
}
\examples{
g_recovered_daily()
g_recovered_daily(saveplot = FALSE, savepng = FALSE)
}
