% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_recovered_growth.R
\name{g_recovered_growth}
\alias{g_recovered_growth}
\title{Recovered growth rate from COVID19}
\usage{
g_recovered_growth(saveplot = FALSE, savepng = FALSE)
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default \code{FALSE}.}

\item{savepng}{Logical. Should save a png version of the plot? Default \code{FALSE}.}
}
\value{
Graph of the recovered persons growth rate and saves a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs the growth rate of those recovered from COVID19
in the Dominican Republic.
}
\examples{
g_recovered_growth(saveplot = FALSE, savepng = TRUE)
}
