% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_recovered_province.R
\name{g_recovered_province}
\alias{g_recovered_province}
\title{Recovered by province from COVID19 in the Dominican Republic}
\usage{
g_recovered_province()
}
\arguments{
\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph of the recovered persons ordered by provinces and saves a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs those recovered from COVID19 in the Dominican Republic ordered by provinces.
}
\examples{
g_recovered_province()
}
