% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_sex.R
\name{g_sex}
\alias{g_sex}
\title{Distribution by sex of COVID19 positives in the Dominican Republic}
\usage{
g_sex()
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default FALSE.}

\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph of the distribution according to sex of the positives and saves a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs the distribution according to the sex of the positives.
}
\examples{
g_sex()
g_sex(saveplot = FALSE, savepng = FALSE)
}
