% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_status.R
\name{g_status}
\alias{g_status}
\title{Distribution by status of affected individuals of COVID19 in the Dominican Republic}
\usage{
g_status()
}
\arguments{
\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph of the distribution according to the status of affected individuals and saves a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs the distribution according to the status of affected individuals. This can be;
1) Home isolation,
2) Hospital isolation,
3) Recovered and
4) Deaths.
}
\examples{
g_status()
}
