% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_tests.R
\name{g_tests}
\alias{g_tests}
\title{COVID19 tests in the Dominican Republic}
\usage{
g_tests(saveplot = FALSE, savepng = FALSE)
}
\arguments{
\item{saveplot}{Logical. Should save the ggplot objet to the \code{.GlobalEnv}? Default \code{FALSE}.}

\item{savepng}{Logical. Should save a png version of the plot? Default \code{FALSE}.}
}
\value{
Graph of daily and accumulated tests both in units and per million inhabitants
and save a copy in png format on the computer to the address defined in \code{setwd()}.
}
\description{
This function graphs the COVID19 tests in RD.
}
\examples{
g_tests()
g_tests(saveplot = FALSE, savepng = TRUE)
}
