% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_world.R
\name{g_world}
\alias{g_world}
\title{Comparison of the COVID19 situation in the Dominican Republic and the world.}
\usage{
g_world()
}
\arguments{
\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
The following scatter graphics are saved in png format on the computer
in the address defined in \code{setwd()}:
1. Lethality against deaths (World and LA)
2. Lethality and positivity (LA)
3. L2 (\code{Deaths / \[Deaths + Recovered\]}) and positive of COVID19 (World and LA)
4. Positives and positivity (World and LA)
5. Positives and recovery rate (World and LA)
}
\description{
This function displays various charts comparing the COVID19 situation
in the Dominican Republic and the world.
}
\examples{
g_world()
}
