% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{load_themes}
\alias{load_themes}
\title{Charts themes}
\usage{
load_themes()
}
\value{
Objects class \code{theme()} containing the themes to use in the charts functions.
}
\description{
This function loads all themes used in the package.
}
\examples{
load_themes()
}
