% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drcovidplots-package.R
\docType{package}
\name{drcovidplots}
\alias{drcovidplots}
\title{drcovidplots}
\description{
Plots for COVID19 data from Dominican Republic
}
\section{drcovidplots functions}{

The drcovidplots functions are: \cr
   \cr
g_cases_province()     \cr
g_death()     \cr
g_density()     \cr
g_evolution_covid()     \cr
g_growth_cases()     \cr
g_hospital()     \cr
g_incidence()     \cr
g_incidence_region()     \cr
g_l2()     \cr
g_lo
g_nc()     \cr
g_np()     \cr
g_np_daily()     \cr
g_positive()     \cr
g_positive_province()     \cr
g_positive_region()     \cr
g_recovered()     \cr
g_recovered_daily()     \cr
g_recovered_growth()     \cr
g_recovered_province()     \cr
g_region()     \cr
g_sex()     \cr
g_status()     \cr
g_tests()     \cr
g_world()     \cr
load_data_covid_dr()     \cr
load_themes()     \cr
}

\section{Warning}{

For all functions to work, you MUST run first, in order, \code{load_data_covid_dr()}
and \code{load_themes()}
}

