% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_region.R
\name{g_region}
\alias{g_region}
\title{COVID19 cases by region}
\usage{
g_region()
}
\arguments{
\item{savepng}{Logical. Should save a png version of the plot? Default FALSE.}
}
\value{
Graph of positive cases of COVID19 by region and save a
copy in png format to the computer at the address defined in \code{setwd()}.
}
\description{
This function graphs the positive cases of COVID19 by region in the DR.
}
\examples{
g_region()
}
