% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data_covid_dr}
\alias{load_data_covid_dr}
\title{Load database from googlesheets and themes for each chart.}
\usage{
load_data_covid_dr()
}
\value{
5 objects class \code{tibble()} containing the information to use in the graphs functions.
Also attached the formats or themes of the graphics to .GlobalEnv.
}
\description{
This function loads information contained in googlessheets about
COVID-19 in the Dominican Republic to .GlobalEnv.
}
\examples{
load_data_covid_dr()
}
